Removed type : NodePort in Step 3
There is no need to configure the Service type as NodePort. In Ingress manifest target type IP is used hence the traffic will be sent directly to the Pod IP. Hence NodePort is unnecessary. Service type ClusterIP is enough. This is a reoccurring misleading  guidance which I requested to be corrected in AWS blogs as well.